+++ common/src/main/java/com/sonicether/soundphysics/SoundPhysics.java
@@ -198,6 +198,13 @@ public class SoundPhysics {
             return null;
         }
 
+        if (!SoundPhysicsMod.CONFIG.updateMovingSounds.get()) {
+            if (category == SoundSource.RECORDS) {
+                setDefaultEnvironment(sourceID);
+                return null;
+            }
+        }
+
         if (SoundPhysicsMod.CONFIG.soundBlacklist.matcher(sound).matches()) {
             setDefaultEnvironment(sourceID);
             return null;